@@ -28,17 +28,17 @@ class CompositeMessageLookup implements MessageLookup {
   ///
   ///  If this locale matches the new one then we can skip looking up the
   ///  messages and assume they will be the same as last time.
-  String _lastLocale;
+  String? _lastLocale;
 
   /// Caches the last messages that we found
-  MessageLookupByLibrary _lastLookup;
+  MessageLookupByLibrary? _lastLookup;
 
   /// Look up the message with the given [name] and [locale] and return the
   /// translated version with the values in [args] interpolated.  If nothing is
   /// found, return the result of [ifAbsent] or [messageText].
-  String lookupMessage(String messageText, String locale, String name,
-      List<dynamic> args, String meaning,
-      {MessageIfAbsent ifAbsent}) {
+  String? lookupMessage(String? messageText, String? locale, String? name,
+      List<Object>? args, String? meaning,
+      {MessageIfAbsent? ifAbsent}) {
     // If passed null, use the default.
     var knownLocale = locale ?? Intl.getCurrentLocale();
     var messages = (knownLocale == _lastLocale)
@@ -54,7 +54,7 @@ class CompositeMessageLookup implements MessageLookup {
   }
 
   /// Find the right message lookup for [locale].
-  MessageLookupByLibrary _lookupMessageCatalog(String locale) {
+  MessageLookupByLibrary? _lookupMessageCatalog(String locale) {
     var verifiedLocale = Intl.verifiedLocale(locale, localeExists,
         onFailure: (locale) => locale);
     _lastLocale = locale;
@@ -97,15 +97,15 @@ abstract class MessageLookupByLibrary {
   /// Ultimately, the information about the enclosing function and its arguments
   /// will be extracted automatically but for the time being it must be passed
   /// explicitly in the [name] and [args] arguments.
-  String lookupMessage(String messageText, String locale, String name,
-      List<dynamic> args, String meaning,
-      {MessageIfAbsent ifAbsent}) {
-    var notFound = false;
+  String? lookupMessage(String? messageText, String? locale, String? name,
+      List<Object>? args, String? meaning,
+      {MessageIfAbsent? ifAbsent}) {
     var actualName = computeMessageName(name, messageText, meaning);
-    if (actualName == null) notFound = true;
-    var translation = this[actualName];
-    notFound = notFound || (translation == null);
-    if (notFound) {
+    Object? translation;
+    if (actualName != null) {
+      translation = this[actualName];
+    }
+    if (translation == null) {
       return ifAbsent == null ? messageText : ifAbsent(messageText, args);
     } else {
       args = args ?? const [];
@@ -114,7 +114,7 @@ abstract class MessageLookupByLibrary {
   }
 
   /// Evaluate the translated message and return the translated string.
-  String evaluateMessage(translation, List<dynamic> args) {
+  String? evaluateMessage(translation, List<dynamic> args) {
     return Function.apply(translation, args);
   }
 
