@@ -2,50 +2,18 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of intl;
-
-// Suppress naming issues as changing them would be breaking.
-// ignore_for_file: constant_identifier_names
-
 /// Bidi stands for Bi-directional text.  According to
 /// http://en.wikipedia.org/wiki/Bi-directional_text: Bi-directional text is
 /// text containing text in both text directionalities, both right-to-left (RTL)
 /// and left-to-right (LTR). It generally involves text containing different
 /// types of alphabets, but may also refer to boustrophedon, which is changing
 /// text directionality in each row.
-///
-/// This file provides some utility classes for determining directionality of
-/// text, switching CSS layout from LTR to RTL, and other normalizing utilities
-/// needed when switching between RTL and LTR formatting.
-///
-/// It defines the TextDirection class which is used to represent directionality
-/// of text,
-/// In most cases, it is preferable to use bidi_formatter.dart, which provides
-/// bidi functionality in the given directional context, instead of using
-/// bidi_utils.dart directly.
-class TextDirection {
-  static const LTR = TextDirection._('LTR', 'ltr');
-  static const RTL = TextDirection._('RTL', 'rtl');
-  // If the directionality of the text cannot be determined and we are not using
-  // the context direction (or if the context direction is unknown), then the
-  // text falls back on the more common ltr direction.
-  static const UNKNOWN = TextDirection._('UNKNOWN', 'ltr');
 
-  /// Textual representation of the directionality constant. One of
-  /// 'LTR', 'RTL', or 'UNKNOWN'.
-  final String value;
+import '../global_state.dart' as global_state;
+import 'text_direction.dart';
 
-  /// Textual representation of the directionality when used in span tag.
-  final String spanText;
-
-  const TextDirection._(this.value, this.spanText);
-
-  /// Returns true if [otherDirection] is known to be different from this
-  /// direction.
-  bool isDirectionChange(TextDirection otherDirection) =>
-      otherDirection != TextDirection.UNKNOWN && this != otherDirection;
-}
-
+// Suppress naming issues as changing them would be breaking.
+// ignore_for_file: constant_identifier_names
 // ignore: avoid_classes_with_only_static_members
 /// This provides utility methods for working with bidirectional text. All
 /// of the methods are static, and are organized into a class primarily to
@@ -144,8 +112,8 @@ class Bidi {
       r'($|-|_)',
       caseSensitive: false);
 
-  static String _lastLocaleCheckedForRtl;
-  static bool _lastRtlCheck;
+  static String? _lastLocaleCheckedForRtl;
+  static bool? _lastRtlCheck;
 
   /// Check if a BCP 47 / III [languageString] indicates an RTL language.
   ///
@@ -168,13 +136,13 @@ class Bidi {
   /// http://www.iana.org/assignments/language-subtag-registry, as well as
   /// Sindhi (sd) and Uyghur (ug).  The presence of other subtags of the
   /// language code, e.g. regions like EG (Egypt), is ignored.
-  static bool isRtlLanguage([String languageString]) {
-    var language = languageString ?? Intl.getCurrentLocale();
+  static bool isRtlLanguage([String? languageString]) {
+    var language = languageString ?? global_state.getCurrentLocale();
     if (_lastLocaleCheckedForRtl != language) {
       _lastLocaleCheckedForRtl = language;
       _lastRtlCheck = _rtlLocaleRegex.hasMatch(language);
     }
-    return _lastRtlCheck;
+    return _lastRtlCheck!;
   }
 
   /// Enforce the [html] snippet in RTL directionality regardless of overall
@@ -216,7 +184,7 @@ class Bidi {
     if (html.startsWith('<')) {
       var buffer = StringBuffer();
       var startIndex = 0;
-      Match match = RegExp('<\\w+').firstMatch(html);
+      var match = RegExp('<\\w+').firstMatch(html);
       if (match != null) {
         buffer
           ..write(html.substring(startIndex, match.end))
@@ -233,7 +201,7 @@ class Bidi {
   /// problem of messy bracket display that frequently happens in RTL layout.
   /// If [isRtlContext] is true, then we explicitly want to wrap in a span of
   /// RTL directionality, regardless of the estimated directionality.
-  static String guardBracketInHtml(String str, [bool isRtlContext]) {
+  static String guardBracketInHtml(String str, [bool? isRtlContext]) {
     var useRtl = isRtlContext == null ? hasAnyRtl(str) : isRtlContext;
     var matchingBrackets =
         RegExp(r'(\(.*?\)+)|(\[.*?\]+)|(\{.*?\}+)|(&lt;.*?(&gt;)+)');
@@ -247,7 +215,7 @@ class Bidi {
   /// as good as guardBracketInHtml. If [isRtlContext] is true, then we
   /// explicitly want to wrap in a span of RTL directionality, regardless of the
   /// estimated directionality.
-  static String guardBracketInText(String str, [bool isRtlContext]) {
+  static String guardBracketInText(String str, [bool? isRtlContext]) {
     var useRtl = isRtlContext == null ? hasAnyRtl(str) : isRtlContext;
     var mark = useRtl ? RLM : LRM;
     return _guardBracketHelper(
@@ -261,7 +229,7 @@ class Bidi {
   /// would return 'firehydrant!'.  // TODO(efortuna): Get rid of this once this
   /// is implemented in Dart.  // See Issue 2979.
   static String _guardBracketHelper(String str, RegExp regexp,
-      [String before, String after]) {
+      [String? before, String? after]) {
     var buffer = StringBuffer();
     var startIndex = 0;
     for (var match in regexp.allMatches(str)) {
