@@ -2,7 +2,17 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of intl;
+import 'package:intl/date_symbols.dart';
+import 'package:intl/src/date_format_internal.dart';
+import 'package:intl/src/intl_helpers.dart' as helpers;
+
+import 'constants.dart' as constants;
+import 'date_builder.dart';
+import 'date_computation.dart' as date_computation;
+import 'intl_stream.dart';
+import 'regexp.dart' as regexp;
+
+part 'date_format_field.dart';
 
 // Suppress naming issues as changes would breaking.
 // ignore_for_file: non_constant_identifier_names, constant_identifier_names
@@ -57,9 +67,6 @@ part of intl;
 /// import 'package:intl/date_symbol_data_http_request.dart';
 /// initializeDateFormatting('pt_BR', null).then((_) => runMyCode());
 /// ```
-///
-/// The code in example/basic/basic_example.dart shows a full example of
-/// using this mechanism.
 ///
 /// Once we have the locale data, we need to specify the particular format.
 /// This library uses the ICU/JDK date/time pattern specification both for
@@ -255,13 +262,13 @@ class DateFormat {
   ///
   /// If [locale] does not exist in our set of supported locales then an
   /// [ArgumentError] is thrown.
-  DateFormat([String newPattern, String locale]) {
+  DateFormat([String? newPattern, String? locale])
+      : _locale = helpers.verifiedLocale(locale, localeExists, null)! {
     // TODO(alanknight): It should be possible to specify multiple skeletons eg
     // date, time, timezone all separately. Adding many or named parameters to
     // the constructor seems awkward, especially with the possibility of
     // confusion with the locale. A 'fluent' interface with cascading on an
     // instance might work better? A list of patterns is also possible.
-    _locale = Intl.verifiedLocale(locale, localeExists);
     addPattern(newPattern);
   }
 
@@ -351,10 +358,9 @@ class DateFormat {
   }
 
   DateTime _parseLoose(String inputString, bool utc) {
-    var dateFields =
-        _DateBuilder(locale ?? Intl.defaultLocale, dateTimeConstructor);
+    var dateFields = DateBuilder(locale, dateTimeConstructor);
     if (utc) dateFields.utc = true;
-    var stream = _Stream(inputString);
+    var stream = IntlStream(inputString);
     for (var field in _formatFields) {
       field.parseLoose(stream, dateFields);
     }
@@ -380,11 +386,10 @@ class DateFormat {
   DateTime _parse(String inputString, {bool utc = false, bool strict = false}) {
     // TODO(alanknight): The Closure code refers to special parsing of numeric
     // values with no delimiters, which we currently don't do. Should we?
-    var dateFields =
-        _DateBuilder(locale ?? Intl.defaultLocale, dateTimeConstructor);
+    var dateFields = DateBuilder(locale, dateTimeConstructor);
     if (utc) dateFields.utc = true;
-    dateFields._dateOnly = dateOnly;
-    var stream = _Stream(inputString);
+    dateFields.dateOnly = dateOnly;
+    var stream = IntlStream(inputString);
     for (var field in _formatFields) {
       field.parse(stream, dateFields);
     }
@@ -400,7 +405,7 @@ class DateFormat {
   ///
   /// For example, 'yyyy-MM-dd' would be true, but 'dd hh:mm' would be false.
   bool get dateOnly => _dateOnly ??= _checkDateOnly;
-  bool _dateOnly;
+  bool? _dateOnly;
   bool get _checkDateOnly => _formatFields.every((each) => each.forDate);
 
   /// Given user input, attempt to parse the [inputString] into the anticipated
@@ -485,15 +490,19 @@ class DateFormat {
   DateFormat.j([locale]) : this('j', locale);
   DateFormat.jm([locale]) : this('jm', locale);
   DateFormat.jms([locale]) : this('jms', locale);
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat.jmv([locale]) : this('jmv', locale);
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat.jmz([locale]) : this('jmz', locale);
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat.jv([locale]) : this('jv', locale);
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat.jz([locale]) : this('jz', locale);
@@ -544,15 +553,19 @@ class DateFormat {
   DateFormat add_j() => addPattern('j');
   DateFormat add_jm() => addPattern('jm');
   DateFormat add_jms() => addPattern('jms');
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat add_jmv() => addPattern('jmv');
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat add_jmz() => addPattern('jmz');
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat add_jv() => addPattern('jv');
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   DateFormat add_jz() => addPattern('jz');
@@ -596,15 +609,19 @@ class DateFormat {
   static const String HOUR = 'j';
   static const String HOUR_MINUTE = 'jm';
   static const String HOUR_MINUTE_SECOND = 'jms';
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   static const String HOUR_MINUTE_GENERIC_TZ = 'jmv';
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   static const String HOUR_MINUTE_TZ = 'jmz';
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   static const String HOUR_GENERIC_TZ = 'jv';
+
   /// NOT YET IMPLEMENTED.
   // TODO(https://github.com/dart-lang/intl/issues/74)
   static const String HOUR_TZ = 'jz';
@@ -618,20 +635,20 @@ class DateFormat {
   /// The full template string. This may have been specified directly, or
   /// it may have been derived from a skeleton and the locale information
   /// on how to interpret that skeleton.
-  String _pattern;
+  String? _pattern;
 
   /// We parse the format string into individual [_DateFormatField] objects
   /// that are used to do the actual formatting and parsing. Do not use
   /// this variable directly, use the getter [_formatFields].
-  List<_DateFormatField> _formatFieldsPrivate;
+  List<_DateFormatField>? _formatFieldsPrivate;
 
   /// Getter for [_formatFieldsPrivate] that lazily initializes it.
   List<_DateFormatField> get _formatFields {
     if (_formatFieldsPrivate == null) {
       if (_pattern == null) _useDefaultPattern();
-      _formatFieldsPrivate = parsePattern(_pattern);
+      _formatFieldsPrivate = parsePattern(_pattern!);
     }
-    return _formatFieldsPrivate;
+    return _formatFieldsPrivate!;
   }
 
   /// We are being asked to do formatting without having set any pattern.
@@ -671,7 +688,7 @@ class DateFormat {
   /// known skeletons.  If it's found there, then use the corresponding pattern
   /// for this locale.  If it's not, then treat [inputPattern] as an explicit
   /// pattern.
-  DateFormat addPattern(String inputPattern, [String separator = ' ']) {
+  DateFormat addPattern(String? inputPattern, [String separator = ' ']) {
     // TODO(alanknight): This is an expensive operation. Caching recently used
     // formats, or possibly introducing an entire 'locale' object that would
     // cache patterns for that locale could be a good optimization.
@@ -687,7 +704,7 @@ class DateFormat {
   }
 
   /// Return the pattern that we use to format dates.
-  String get pattern => _pattern;
+  String? get pattern => _pattern;
 
   /// Return the skeletons for our current locale.
   Map<dynamic, dynamic> get _availableSkeletons => dateTimePatterns[locale];
@@ -696,14 +713,15 @@ class DateFormat {
   ///
   /// This can be useful to find lists like the names of weekdays or months in a
   /// locale, but the structure of this data may change, and it's generally
-  /// better to go through the [format] and [parse] APIs. If the locale isn't
-  /// present, or is uninitialized, returns null.
+  /// better to go through the [format] and [parse] APIs.
+  ///
+  /// If the locale isn't present, or is uninitialized, throws.
   DateSymbols get dateSymbols {
     if (_locale != lastDateSymbolLocale) {
       lastDateSymbolLocale = _locale;
       cachedDateSymbols = dateTimeSymbols[_locale];
     }
-    return cachedDateSymbols;
+    return cachedDateSymbols!;
   }
 
   static final Map<String, bool> _useNativeDigitsByDefault = {};
@@ -730,14 +748,14 @@ class DateFormat {
     _useNativeDigitsByDefault[locale] = value;
   }
 
-  bool _useNativeDigits;
+  bool? _useNativeDigits;
 
   /// Should we use native digits for printing DateTime, or ASCII.
   ///
   /// The default for this can be set using [useNativeDigitsByDefaultFor].
   bool get useNativeDigits => _useNativeDigits == null
       ? _useNativeDigits = shouldUseNativeDigitsByDefaultFor(locale)
-      : _useNativeDigits;
+      : _useNativeDigits!;
 
   /// Should we use native digits for printing DateTime, or ASCII.
   set useNativeDigits(bool native) {
@@ -755,36 +773,32 @@ class DateFormat {
   /// locale.
   static final Map<String, RegExp> _digitMatchers = {};
 
-  RegExp _digitMatcher;
+  RegExp? _digitMatcher;
 
   /// A regular expression which matches against digits for this locale.
   RegExp get digitMatcher {
-    if (_digitMatcher != null) return _digitMatcher;
+    if (_digitMatcher != null) return _digitMatcher!;
     _digitMatcher = _digitMatchers.putIfAbsent(localeZero, _initDigitMatcher);
-    return _digitMatcher;
+    return _digitMatcher!;
   }
 
-  /// Hard-code the most common matcher, which has special RegExp syntax.
-  static final RegExp _asciiDigitMatcher = RegExp(r'^\d+');
-
-  int _localeZeroCodeUnit;
+  int? _localeZeroCodeUnit;
 
   /// For performance, keep the code unit of the zero digit available.
   int get localeZeroCodeUnit => _localeZeroCodeUnit == null
       ? _localeZeroCodeUnit = localeZero.codeUnitAt(0)
-      : _localeZeroCodeUnit;
-  static final int _asciiZeroCodeUnit = '0'.codeUnitAt(0);
+      : _localeZeroCodeUnit!;
 
-  String _localeZero;
+  String? _localeZero;
 
   /// For performance, keep the zero digit available.
   String get localeZero => _localeZero == null
       ? _localeZero = useNativeDigits ? dateSymbols.ZERODIGIT ?? '0' : '0'
-      : _localeZero;
+      : _localeZero!;
 
   // Does this use non-ASCII digits, e.g. Eastern Arabic.
   bool get usesNativeDigits =>
-      useNativeDigits && _localeZeroCodeUnit != _asciiZeroCodeUnit;
+      useNativeDigits && _localeZeroCodeUnit != constants.asciiZeroCodeUnit;
 
   /// Does this use ASCII digits
   bool get usesAsciiDigits => !usesNativeDigits;
@@ -793,10 +807,11 @@ class DateFormat {
   /// locale digits.
   String _localizeDigits(String numberString) {
     if (usesAsciiDigits) return numberString;
-    var newDigits = List<int>(numberString.length);
+    var newDigits = List<int>.filled(numberString.length, 0);
     var oldDigits = numberString.codeUnits;
     for (var i = 0; i < numberString.length; i++) {
-      newDigits[i] = oldDigits[i] + localeZeroCodeUnit - _asciiZeroCodeUnit;
+      newDigits[i] =
+          oldDigits[i] + localeZeroCodeUnit - constants.asciiZeroCodeUnit;
     }
     return String.fromCharCodes(newDigits);
   }
@@ -804,7 +819,7 @@ class DateFormat {
   /// A regular expression that matches for digits in a particular
   /// locale, defined by the digit for zero in that locale.
   RegExp _initDigitMatcher() {
-    if (usesAsciiDigits) return _asciiDigitMatcher;
+    if (usesAsciiDigits) return regexp.asciiDigitMatcher;
     var localeDigits = Iterable.generate(10, (i) => i)
         .map((i) => localeZeroCodeUnit + i)
         .toList();
@@ -828,7 +843,6 @@ class DateFormat {
 
   /// Parse the template pattern and return a list of field objects.
   List<_DateFormatField> parsePattern(String pattern) {
-    if (pattern == null) return null;
     return _parsePatternHelper(pattern).reversed.toList();
   }
 
@@ -846,12 +860,12 @@ class DateFormat {
   }
 
   /// Find elements in a string that are patterns for specific fields.
-  _DateFormatField _match(String pattern) {
+  _DateFormatField? _match(String pattern) {
     for (var i = 0; i < _matchers.length; i++) {
       var regex = _matchers[i];
       var match = regex.firstMatch(pattern);
       if (match != null) {
-        return _fieldConstructors[i](match.group(0), this);
+        return _fieldConstructors[i](match.group(0)!, this);
       }
     }
     return null;
