@@ -2,13 +2,23 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of intl;
+import 'dart:math';
+
+import 'package:intl/number_symbols.dart';
+import 'package:intl/number_symbols_data.dart';
+import 'package:intl/src/intl_helpers.dart' as helpers;
+
+import 'constants.dart' as constants;
+import 'number_format_parser.dart';
+import 'number_parser.dart';
+
+part 'compact_number_format.dart';
 
 // ignore_for_file: constant_identifier_names
 
 /// The function that we pass internally to NumberFormat to get
 /// the appropriate pattern (e.g. currency)
-typedef _PatternGetter = String Function(NumberSymbols);
+typedef _PatternGetter = String? Function(NumberSymbols);
 
 /// Provides the ability to format a number in a locale-specific way.
 ///
@@ -50,49 +60,44 @@ typedef _PatternGetter = String Function(NumberSymbols);
 /// equivalent to "#E0" and does not take into account significant digits.
 class NumberFormat {
   /// Variables to determine how number printing behaves.
-  // TODO(alanknight): If these remain as variables and are set based on the
-  // pattern, can we make them final?
-  String _negativePrefix = '-';
-  String _positivePrefix = '';
-  String _negativeSuffix = '';
-  String _positiveSuffix = '';
+  final String negativePrefix;
+  final String positivePrefix;
+  final String negativeSuffix;
+  final String positiveSuffix;
 
   /// How many numbers in a group when using punctuation to group digits in
   /// large numbers. e.g. in en_US: "1,000,000" has a grouping size of 3 digits
   /// between commas.
-  int _groupingSize = 3;
+  int _groupingSize;
 
   /// In some formats the last grouping size may be different than previous
   /// ones, e.g. Hindi.
-  int _finalGroupingSize = 3;
+  int _finalGroupingSize;
 
   /// Set to true if the format has explicitly set the grouping size.
-  bool _groupingSizeSetExplicitly = false;
-  bool _decimalSeparatorAlwaysShown = false;
-  bool _useSignForPositiveExponent = false;
-  bool _useExponentialNotation = false;
+  final bool _decimalSeparatorAlwaysShown;
+  final bool _useSignForPositiveExponent;
+  final bool _useExponentialNotation;
 
   /// Explicitly store if we are a currency format, and so should use the
   /// appropriate number of decimal digits for a currency.
   // TODO(alanknight): Handle currency formats which are specified in a raw
   /// pattern, not using one of the currency constructors.
-  bool _isForCurrency = false;
+  final bool _isForCurrency;
 
-  int maximumIntegerDigits = 40;
-  int minimumIntegerDigits = 1;
-  int maximumFractionDigits = 3;
-  int minimumFractionDigits = 0;
-  int minimumExponentDigits = 0;
-  int _significantDigits = 0;
-
-  static final _ln10 = log(10);
+  int maximumIntegerDigits;
+  int minimumIntegerDigits;
+  int maximumFractionDigits;
+  int minimumFractionDigits;
+  int minimumExponentDigits;
+  int? _significantDigits;
 
   ///  How many significant digits should we print.
   ///
   ///  Note that if significantDigitsInUse is the default false, this
   ///  will be ignored.
-  int get significantDigits => _significantDigits;
-  set significantDigits(int x) {
+  int? get significantDigits => _significantDigits;
+  set significantDigits(int? x) {
     _significantDigits = x;
     significantDigitsInUse = true;
   }
@@ -101,39 +106,28 @@ class NumberFormat {
 
   /// For percent and permille, what are we multiplying by in order to
   /// get the printed value, e.g. 100 for percent.
-  int get _multiplier => _internalMultiplier;
-  set _multiplier(int x) {
-    _internalMultiplier = x;
-    _multiplierDigits = (log(_multiplier) / _ln10).round();
-  }
+  final int multiplier;
 
-  int _internalMultiplier = 1;
-
-  /// How many digits are there in the [_multiplier].
-  int _multiplierDigits = 0;
+  /// How many digits are there in the [multiplier].
+  final int _multiplierDigits;
 
   /// Stores the pattern used to create this format. This isn't used, but
   /// is helpful in debugging.
-  String _pattern;
+  final String? _pattern;
 
   /// The locale in which we print numbers.
   final String _locale;
 
   /// Caches the symbols used for our locale.
-  NumberSymbols _symbols;
+  final NumberSymbols _symbols;
 
   /// The name of the currency to print, in ISO 4217 form.
-  String currencyName;
+  String? currencyName;
 
   /// The symbol to be used when formatting this as currency.
   ///
   /// For example, "$", "US$", or "€".
-  String _currencySymbol;
-
-  /// The symbol to be used when formatting this as currency.
-  ///
-  /// For example, "$", "US$", or "€".
-  String get currencySymbol => _currencySymbol ?? currencyName;
+  final String currencySymbol;
 
   /// The number of decimal places to use when formatting.
   ///
@@ -150,20 +144,7 @@ class NumberFormat {
   ///       NumberFormat.currency(locale: 'en_US')
   /// will format with two, which is the default for that locale.
   ///
-  int get decimalDigits => _decimalDigits;
-
-  int _decimalDigits;
-
-  /// For currencies, the default number of decimal places to use in
-  /// formatting. Defaults to two for non-currencies or currencies where it's
-  /// not specified.
-  int get _defaultDecimalDigits =>
-      currencyFractionDigits[currencyName.toUpperCase()] ??
-      currencyFractionDigits['DEFAULT'];
-
-  /// If we have a currencyName, use the decimal digits for that currency,
-  /// unless we've explicitly specified some other number.
-  bool get _overridesDecimalDigits => decimalDigits != null || _isForCurrency;
+  final int? decimalDigits;
 
   /// Transient internal state in which to build up the result of the format
   /// operation. We can have this be just an instance variable because Dart is
@@ -174,25 +155,26 @@ class NumberFormat {
 
   /// Create a number format that prints using [newPattern] as it applies in
   /// [locale].
-  factory NumberFormat([String newPattern, String locale]) =>
+  factory NumberFormat([String? newPattern, String? locale]) =>
       NumberFormat._forPattern(locale, (x) => newPattern);
 
   /// Create a number format that prints as DECIMAL_PATTERN.
-  NumberFormat.decimalPattern([String locale])
-      : this._forPattern(locale, (x) => x.DECIMAL_PATTERN);
+  factory NumberFormat.decimalPattern([String? locale]) =>
+      NumberFormat._forPattern(locale, (x) => x.DECIMAL_PATTERN);
 
   /// Create a number format that prints as PERCENT_PATTERN.
-  NumberFormat.percentPattern([String locale])
-      : this._forPattern(locale, (x) => x.PERCENT_PATTERN);
+  factory NumberFormat.percentPattern([String? locale]) =>
+      NumberFormat._forPattern(locale, (x) => x.PERCENT_PATTERN);
 
   /// Create a number format that prints as PERCENT_PATTERN.
-  NumberFormat.decimalPercentPattern({String locale, int decimalDigits})
-      : this._forPattern(locale, (x) => x.PERCENT_PATTERN,
-            decimalDigits: decimalDigits);
+  factory NumberFormat.decimalPercentPattern(
+          {String? locale, int? decimalDigits}) =>
+      NumberFormat._forPattern(locale, (x) => x.PERCENT_PATTERN,
+          decimalDigits: decimalDigits);
 
   /// Create a number format that prints as SCIENTIFIC_PATTERN.
-  NumberFormat.scientificPattern([String locale])
-      : this._forPattern(locale, (x) => x.SCIENTIFIC_PATTERN);
+  factory NumberFormat.scientificPattern([String? locale]) =>
+      NumberFormat._forPattern(locale, (x) => x.SCIENTIFIC_PATTERN);
 
   /// A regular expression to validate currency names are exactly three
   /// alphabetic characters.
@@ -207,7 +189,7 @@ class NumberFormat {
   ///           .currencyPattern(Intl.defaultLocale, "€");
   @Deprecated('Use NumberFormat.currency')
   factory NumberFormat.currencyPattern(
-      [String locale, String currencyNameOrSymbol]) {
+      [String? locale, String? currencyNameOrSymbol]) {
     // If it looks like an iso4217 name, pass as name, otherwise as symbol.
     if (currencyNameOrSymbol != null &&
         _checkCurrencyName.hasMatch(currencyNameOrSymbol)) {
@@ -254,17 +236,18 @@ class NumberFormat {
   /// format. This is useful if you have your own locale data which includes
   /// unsupported formats (e.g. accounting format for currencies.)
   // TODO(alanknight): Should we allow decimalDigits on other numbers.
-  NumberFormat.currency(
-      {String locale,
-      String name,
-      String symbol,
-      int decimalDigits,
-      String customPattern})
-      : this._forPattern(locale, (x) => customPattern ?? x.CURRENCY_PATTERN,
-            name: name,
-            currencySymbol: symbol,
-            decimalDigits: decimalDigits,
-            isForCurrency: true);
+  factory NumberFormat.currency(
+          {String? locale,
+          String? name,
+          String? symbol,
+          int? decimalDigits,
+          String? customPattern}) =>
+      NumberFormat._forPattern(
+          locale, (x) => customPattern ?? x.CURRENCY_PATTERN,
+          name: name,
+          currencySymbol: symbol,
+          decimalDigits: decimalDigits,
+          isForCurrency: true);
 
   /// Creates a [NumberFormat] for currencies, using the simple symbol for the
   /// currency if one is available (e.g. $, €), so it should only be used if the
@@ -290,12 +273,11 @@ class NumberFormat {
   ///       NumberFormat.simpleCurrency(locale: 'en_US')
   /// will format with two, which is the default for that locale.
   factory NumberFormat.simpleCurrency(
-      {String locale, String name, int decimalDigits}) {
+      {String? locale, String? name, int? decimalDigits}) {
     return NumberFormat._forPattern(locale, (x) => x.CURRENCY_PATTERN,
         name: name,
-        computeCurrencySymbol: (format) =>
-            _simpleCurrencySymbols[format.currencyName] ?? format.currencyName,
         decimalDigits: decimalDigits,
+        lookupSimpleCurrencySymbol: true,
         isForCurrency: true);
   }
 
@@ -313,182 +295,7 @@ class NumberFormat {
   /// (The current implementation is the same for all locales, but this is
   /// temporary and callers shouldn't rely on it.)
   String simpleCurrencySymbol(String currencyCode) =>
-      _simpleCurrencySymbols[currencyCode] ?? currencyCode;
-
-  /// A map from currency names to the simple name/symbol.
-  ///
-  /// The simple currency symbol is generally short, and the same or related to
-  /// what is used in countries having the currency as an official symbol. It
-  /// may be a symbol character, or may have letters, or both. It may be
-  /// different according to the locale: for example, for an Arabic locale it
-  /// may consist of Arabic letters, but for a French locale consist of Latin
-  /// letters. It will not be unique: for example, "$" can appear for both USD
-  /// and CAD.
-  ///
-  /// (The current implementation is the same for all locales, but this is
-  /// temporary and callers shouldn't rely on it.)
-  static final Map<String, String> _simpleCurrencySymbols = {
-    'AFN': 'Af.',
-    'TOP': r'T$',
-    'MGA': 'Ar',
-    'THB': '\u0e3f',
-    'PAB': 'B/.',
-    'ETB': 'Birr',
-    'VEF': 'Bs',
-    'BOB': 'Bs',
-    'GHS': 'GHS',
-    'CRC': '\u20a1',
-    'NIO': r'C$',
-    'GMD': 'GMD',
-    'MKD': 'din',
-    'BHD': 'din',
-    'DZD': 'din',
-    'IQD': 'din',
-    'JOD': 'din',
-    'KWD': 'din',
-    'LYD': 'din',
-    'RSD': 'din',
-    'TND': 'din',
-    'AED': 'dh',
-    'MAD': 'dh',
-    'STD': 'Db',
-    'BSD': r'$',
-    'FJD': r'$',
-    'GYD': r'$',
-    'KYD': r'$',
-    'LRD': r'$',
-    'SBD': r'$',
-    'SRD': r'$',
-    'AUD': r'$',
-    'BBD': r'$',
-    'BMD': r'$',
-    'BND': r'$',
-    'BZD': r'$',
-    'CAD': r'$',
-    'HKD': r'$',
-    'JMD': r'$',
-    'NAD': r'$',
-    'NZD': r'$',
-    'SGD': r'$',
-    'TTD': r'$',
-    'TWD': r'NT$',
-    'USD': r'$',
-    'XCD': r'$',
-    'VND': '\u20ab',
-    'AMD': 'Dram',
-    'CVE': 'CVE',
-    'EUR': '\u20ac',
-    'AWG': 'Afl.',
-    'HUF': 'Ft',
-    'BIF': 'FBu',
-    'CDF': 'FrCD',
-    'CHF': 'CHF',
-    'DJF': 'Fdj',
-    'GNF': 'FG',
-    'RWF': 'RF',
-    'XOF': 'CFA',
-    'XPF': 'FCFP',
-    'KMF': 'CF',
-    'XAF': 'FCFA',
-    'HTG': 'HTG',
-    'PYG': 'Gs',
-    'UAH': '\u20b4',
-    'PGK': 'PGK',
-    'LAK': '\u20ad',
-    'CZK': 'K\u010d',
-    'SEK': 'kr',
-    'ISK': 'kr',
-    'DKK': 'kr',
-    'NOK': 'kr',
-    'HRK': 'kn',
-    'MWK': 'MWK',
-    'ZMK': 'ZWK',
-    'AOA': 'Kz',
-    'MMK': 'K',
-    'GEL': 'GEL',
-    'LVL': 'Ls',
-    'ALL': 'Lek',
-    'HNL': 'L',
-    'SLL': 'SLL',
-    'MDL': 'MDL',
-    'RON': 'RON',
-    'BGN': 'lev',
-    'SZL': 'SZL',
-    'TRY': 'TL',
-    'LTL': 'Lt',
-    'LSL': 'LSL',
-    'AZN': 'man.',
-    'BAM': 'KM',
-    'MZN': 'MTn',
-    'NGN': '\u20a6',
-    'ERN': 'Nfk',
-    'BTN': 'Nu.',
-    'MRO': 'MRO',
-    'MOP': 'MOP',
-    'CUP': r'$',
-    'CUC': r'$',
-    'ARS': r'$',
-    'CLF': 'UF',
-    'CLP': r'$',
-    'COP': r'$',
-    'DOP': r'$',
-    'MXN': r'$',
-    'PHP': '\u20b1',
-    'UYU': r'$',
-    'FKP': '£',
-    'GIP': '£',
-    'SHP': '£',
-    'EGP': 'E£',
-    'LBP': 'L£',
-    'SDG': 'SDG',
-    'SSP': 'SSP',
-    'GBP': '£',
-    'SYP': '£',
-    'BWP': 'P',
-    'GTQ': 'Q',
-    'ZAR': 'R',
-    'BRL': r'R$',
-    'OMR': 'Rial',
-    'QAR': 'Rial',
-    'YER': 'Rial',
-    'IRR': 'Rial',
-    'KHR': 'Riel',
-    'MYR': 'RM',
-    'SAR': 'Riyal',
-    'BYR': 'BYR',
-    'RUB': 'руб.',
-    'MUR': 'Rs',
-    'SCR': 'SCR',
-    'LKR': 'Rs',
-    'NPR': 'Rs',
-    'INR': '\u20b9',
-    'PKR': 'Rs',
-    'IDR': 'Rp',
-    'ILS': '\u20aa',
-    'KES': 'Ksh',
-    'SOS': 'SOS',
-    'TZS': 'TSh',
-    'UGX': 'UGX',
-    'PEN': 'S/.',
-    'KGS': 'KGS',
-    'UZS': 'so\u02bcm',
-    'TJS': 'Som',
-    'BDT': '\u09f3',
-    'WST': 'WST',
-    'KZT': '\u20b8',
-    'MNT': '\u20ae',
-    'VUV': 'VUV',
-    'KPW': '\u20a9',
-    'KRW': '\u20a9',
-    'JPY': '¥',
-    'CNY': '¥',
-    'PLN': 'z\u0142',
-    'MVR': 'Rf',
-    'NLG': 'NAf',
-    'ZMW': 'ZK',
-    'ANG': 'ƒ',
-    'TMT': 'TMT',
-  };
+      constants.simpleCurrencySymbols[currencyCode] ?? currencyCode;
 
   /// Create a number format that prints in a pattern we get from
   /// the [getPattern] function using the locale [locale].
@@ -496,30 +303,68 @@ class NumberFormat {
   /// The [currencySymbol] can either be specified directly, or we can pass a
   /// function [computeCurrencySymbol] that will compute it later, given other
   /// information, typically the verified locale.
-  NumberFormat._forPattern(String locale, _PatternGetter getPattern,
-      {String name,
-      String currencySymbol,
-      String Function(NumberFormat) computeCurrencySymbol,
-      int decimalDigits,
-      bool isForCurrency = false})
-      : _locale = Intl.verifiedLocale(locale, localeExists),
-        _isForCurrency = isForCurrency {
-    _currencySymbol = currencySymbol;
-    _decimalDigits = decimalDigits;
-    _symbols = numberFormatSymbols[_locale];
-    _localeZero = _symbols.ZERO_DIGIT.codeUnitAt(0);
-    _zeroOffset = _localeZero - _zero;
-    _negativePrefix = _symbols.MINUS_SIGN;
-    currencyName = name ?? _symbols.DEF_CURRENCY_CODE;
-    if (_currencySymbol == null && computeCurrencySymbol != null) {
-      _currencySymbol = computeCurrencySymbol(this);
+  factory NumberFormat._forPattern(String? locale, _PatternGetter getPattern,
+      {String? name,
+      String? currencySymbol,
+      int? decimalDigits,
+      bool lookupSimpleCurrencySymbol = false,
+      bool isForCurrency = false}) {
+    locale = helpers.verifiedLocale(locale, localeExists, null)!;
+    var symbols = numberFormatSymbols[locale] as NumberSymbols;
+    var localeZero = symbols.ZERO_DIGIT.codeUnitAt(0);
+    var zeroOffset = localeZero - constants.asciiZeroCodeUnit;
+    name ??= symbols.DEF_CURRENCY_CODE;
+    if (currencySymbol == null && lookupSimpleCurrencySymbol) {
+      currencySymbol = constants.simpleCurrencySymbols[name];
     }
-    _setPattern(getPattern(_symbols));
+    currencySymbol ??= name;
+
+    var pattern = getPattern(symbols);
+
+    return NumberFormat._(
+        name,
+        currencySymbol,
+        isForCurrency,
+        locale,
+        localeZero,
+        pattern,
+        symbols,
+        zeroOffset,
+        NumberFormatParser.parse(symbols, pattern, isForCurrency,
+            currencySymbol, name, decimalDigits));
   }
+
+  NumberFormat._(
+      this.currencyName,
+      this.currencySymbol,
+      this._isForCurrency,
+      this._locale,
+      this.localeZero,
+      this._pattern,
+      this._symbols,
+      this._zeroOffset,
+      NumberFormatParseResult result)
+      : positivePrefix = result.positivePrefix,
+        negativePrefix = result.negativePrefix,
+        positiveSuffix = result.positiveSuffix,
+        negativeSuffix = result.negativeSuffix,
+        multiplier = result.multiplier,
+        _multiplierDigits = result.multiplierDigits,
+        _useExponentialNotation = result.useExponentialNotation,
+        minimumExponentDigits = result.minimumExponentDigits,
+        maximumIntegerDigits = result.maximumIntegerDigits,
+        minimumIntegerDigits = result.minimumIntegerDigits,
+        maximumFractionDigits = result.maximumFractionDigits,
+        minimumFractionDigits = result.minimumFractionDigits,
+        _groupingSize = result.groupingSize,
+        _finalGroupingSize = result.finalGroupingSize,
+        _useSignForPositiveExponent = result.useSignForPositiveExponent,
+        _decimalSeparatorAlwaysShown = result.decimalSeparatorAlwaysShown,
+        decimalDigits = result.decimalDigits;
 
   /// A number format for compact representations, e.g. "1.2M" instead
   /// of "1,200,000".
-  factory NumberFormat.compact({String locale}) {
+  factory NumberFormat.compact({String? locale}) {
     return _CompactNumberFormat(
         locale: locale,
         formatType: _CompactFormatType.COMPACT_DECIMAL_SHORT_PATTERN);
@@ -527,7 +372,7 @@ class NumberFormat {
 
   /// A number format for "long" compact representations, e.g. "1.2 million"
   /// instead of of "1,200,000".
-  factory NumberFormat.compactLong({String locale}) {
+  factory NumberFormat.compactLong({String? locale}) {
     return _CompactNumberFormat(
         locale: locale,
         formatType: _CompactFormatType.COMPACT_DECIMAL_LONG_PATTERN);
@@ -538,22 +383,21 @@ class NumberFormat {
   /// based on the currency name or the locale. See
   /// [NumberFormat.simpleCurrency].
   factory NumberFormat.compactSimpleCurrency(
-      {String locale, String name, int decimalDigits}) {
+      {String? locale, String? name, int? decimalDigits}) {
     return _CompactNumberFormat(
         locale: locale,
         formatType: _CompactFormatType.COMPACT_DECIMAL_SHORT_CURRENCY_PATTERN,
         name: name,
         getPattern: (symbols) => symbols.CURRENCY_PATTERN,
-        computeCurrencySymbol: (format) =>
-            _simpleCurrencySymbols[format.currencyName] ?? format.currencyName,
         decimalDigits: decimalDigits,
+        lookupSimpleCurrencySymbol: true,
         isForCurrency: true);
   }
 
   /// A number format for compact currency representations, e.g. "$1.2M" instead
   /// of "$1,200,000".
   factory NumberFormat.compactCurrency(
-      {String locale, String name, String symbol, int decimalDigits}) {
+      {String? locale, String? name, String? symbol, int? decimalDigits}) {
     return _CompactNumberFormat(
         locale: locale,
         formatType: _CompactFormatType.COMPACT_DECIMAL_SHORT_CURRENCY_PATTERN,
@@ -594,7 +438,7 @@ class NumberFormat {
 
   /// Parse the number represented by the string. If it's not
   /// parseable, throws a [FormatException].
-  num parse(String text) => _NumberParser(this, text).value;
+  num parse(String text) => NumberParser(this, text).value!;
 
   /// Format the main part of the number in the form dictated by the pattern.
   void _formatNumber(number) {
@@ -759,6 +603,7 @@ class NumberFormat {
       /// If we have significant digits, recalculate the number of fraction
       /// digits based on that.
       if (significantDigitsInUse) {
+        var significantDigits = this.significantDigits!;
         var integerLength = numberOfIntegerDigits(integerPart);
         var remainingSignificantDigits =
             significantDigits - _multiplierDigits - integerLength;
@@ -769,8 +614,8 @@ class NumberFormat {
           integerPart = (integerPart / divideBy).round() * divideBy;
         }
       }
-      power = pow(10, fractionDigits);
-      digitMultiplier = power * _multiplier;
+      power = pow(10, fractionDigits) as int;
+      digitMultiplier = power * multiplier;
 
       // Multiply out to the number of decimal places and the percent, then
       // round. For fixed-size integer types this should always be zero, so
@@ -839,9 +684,9 @@ class NumberFormat {
   String _mainIntegerDigits(integer) {
     if (integer == 0) return '';
     var digits = integer.toString();
-    if (significantDigitsInUse && digits.length > significantDigits) {
+    if (significantDigitsInUse && digits.length > significantDigits!) {
       digits = digits.substring(0, significantDigits) +
-          ''.padLeft(digits.length - significantDigits, '0');
+          ''.padLeft(digits.length - significantDigits!, '0');
     }
     // If we have a fixed-length int representation, it can have a negative
     // number whose negation is also negative, e.g. 2^-63 in 64-bit.
@@ -852,7 +697,8 @@ class NumberFormat {
   /// Format the part after the decimal place in a fixed point number.
   void _formatFractionPart(String fractionPart) {
     var fractionLength = fractionPart.length;
-    while (fractionPart.codeUnitAt(fractionLength - 1) == _zero &&
+    while (fractionPart.codeUnitAt(fractionLength - 1) ==
+            constants.asciiZeroCodeUnit &&
         fractionLength > minimumFractionDigits + 1) {
       fractionLength--;
     }
@@ -924,41 +770,24 @@ class NumberFormat {
     }
   }
 
-  /// The code point for the character '0'.
-  static const _zero = 48;
-
   /// The code point for the locale's zero digit.
   ///
   ///  Initialized when the locale is set.
-  int _localeZero = 0;
+  final int localeZero;
 
   /// The difference between our zero and '0'.
   ///
   /// In other words, a constant _localeZero - _zero. Initialized when
   /// the locale is set.
-  int _zeroOffset = 0;
+  final int _zeroOffset;
 
   /// Returns the prefix for [x] based on whether it's positive or negative.
   /// In en_US this would be '' and '-' respectively.
-  String _signPrefix(x) => x.isNegative ? _negativePrefix : _positivePrefix;
+  String _signPrefix(x) => x.isNegative ? negativePrefix : positivePrefix;
 
   /// Returns the suffix for [x] based on wether it's positive or negative.
   /// In en_US there are no suffixes for positive or negative.
-  String _signSuffix(x) => x.isNegative ? _negativeSuffix : _positiveSuffix;
-
-  void _setPattern(String newPattern) {
-    if (newPattern == null) return;
-    // Make spaces non-breaking
-    _pattern = newPattern.replaceAll(' ', '\u00a0');
-    var parser =
-        _NumberFormatParser(this, newPattern, currencySymbol, decimalDigits);
-    parser.parse();
-    if (_overridesDecimalDigits) {
-      _decimalDigits ??= _defaultDecimalDigits;
-      minimumFractionDigits = _decimalDigits;
-      maximumFractionDigits = _decimalDigits;
-    }
-  }
+  String _signSuffix(x) => x.isNegative ? negativeSuffix : positiveSuffix;
 
   /// Explicitly turn off any grouping (e.g. by thousands) in this format.
   ///
@@ -972,645 +801,4 @@ class NumberFormat {
   String toString() => 'NumberFormat($_locale, $_pattern)';
 }
 
-///  A one-time object for parsing a particular numeric string. One-time here
-/// means an instance can only parse one string. This is implemented by
-/// transforming from a locale-specific format to one that the system can parse,
-/// then calls the system parsing methods on it.
-class _NumberParser {
-  /// The format for which we are parsing.
-  final NumberFormat format;
-
-  /// The text we are parsing.
-  final String text;
-
-  /// What we use to iterate over the input text.
-  final _Stream input;
-
-  /// The result of parsing [text] according to [format]. Automatically
-  /// populated in the constructor.
-  num value;
-
-  /// The symbols used by our format.
-  NumberSymbols get symbols => format.symbols;
-
-  /// Where we accumulate the normalized representation of the number.
-  final StringBuffer _normalized = StringBuffer();
-
-  /// Did we see something that indicates this is, or at least might be,
-  /// a positive number.
-  bool gotPositive = false;
-
-  /// Did we see something that indicates this is, or at least might be,
-  /// a negative number.
-  bool gotNegative = false;
-
-  /// Did we see the required positive suffix at the end. Should
-  /// match [gotPositive].
-  bool gotPositiveSuffix = false;
-
-  /// Did we see the required negative suffix at the end. Should
-  /// match [gotNegative].
-  bool gotNegativeSuffix = false;
-
-  /// Should we stop parsing before hitting the end of the string.
-  bool done = false;
-
-  /// Have we already skipped over any required prefixes.
-  bool prefixesSkipped = false;
-
-  /// If the number is percent or permill, what do we divide by at the end.
-  int scale = 1;
-
-  String get _positivePrefix => format._positivePrefix;
-  String get _negativePrefix => format._negativePrefix;
-  String get _positiveSuffix => format._positiveSuffix;
-  String get _negativeSuffix => format._negativeSuffix;
-  int get _zero => NumberFormat._zero;
-  int get _localeZero => format._localeZero;
-
-  ///  Create a new [_NumberParser] on which we can call parse().
-  _NumberParser(this.format, this.text) : input = _Stream(text) {
-    scale = format._internalMultiplier;
-    value = parse();
-  }
-
-  ///  The strings we might replace with functions that return the replacement
-  /// values. They are functions because we might need to check something
-  /// in the context. Note that the ordering is important here. For example,
-  /// `symbols.PERCENT` might be " %", and we must handle that before we
-  /// look at an individual space.
-  Map<String, Function> get replacements =>
-      _replacements ??= _initializeReplacements();
-
-  Map<String, Function> _replacements;
-
-  Map<String, Function> _initializeReplacements() => {
-        symbols.DECIMAL_SEP: () => '.',
-        symbols.EXP_SYMBOL: () => 'E',
-        symbols.GROUP_SEP: handleSpace,
-        symbols.PERCENT: () {
-          scale = _NumberFormatParser._PERCENT_SCALE;
-          return '';
-        },
-        symbols.PERMILL: () {
-          scale = _NumberFormatParser._PER_MILLE_SCALE;
-          return '';
-        },
-        ' ': handleSpace,
-        '\u00a0': handleSpace,
-        '+': () => '+',
-        '-': () => '-',
-      };
-
-  void invalidFormat() =>
-      throw FormatException('Invalid number: ${input.contents}');
-
-  /// Replace a space in the number with the normalized form. If space is not
-  /// a significant character (normally grouping) then it's just invalid. If it
-  /// is the grouping character, then it's only valid if it's followed by a
-  /// digit. e.g. '$12 345.00'
-  void handleSpace() =>
-      groupingIsNotASpaceOrElseItIsSpaceFollowedByADigit ? '' : invalidFormat();
-
-  /// Determine if a space is a valid character in the number. See
-  /// [handleSpace].
-  bool get groupingIsNotASpaceOrElseItIsSpaceFollowedByADigit {
-    if (symbols.GROUP_SEP != '\u00a0' || symbols.GROUP_SEP != ' ') return true;
-    var peeked = input.peek(symbols.GROUP_SEP.length + 1);
-    return asDigit(peeked[peeked.length - 1]) != null;
-  }
-
-  /// Turn [char] into a number representing a digit, or null if it doesn't
-  /// represent a digit in this locale.
-  int asDigit(String char) {
-    var charCode = char.codeUnitAt(0);
-    var digitValue = charCode - _localeZero;
-    if (digitValue >= 0 && digitValue < 10) {
-      return digitValue;
-    } else {
-      return null;
-    }
-  }
-
-  /// Check to see if the input begins with either the positive or negative
-  /// prefixes. Set the [gotPositive] and [gotNegative] variables accordingly.
-  void checkPrefixes({bool skip = false}) {
-    bool checkPrefix(String prefix) =>
-        prefix.isNotEmpty && input.startsWith(prefix);
-
-    // TODO(alanknight): There's a faint possibility of a bug here where
-    // a positive prefix is followed by a negative prefix that's also a valid
-    // part of the number, but that seems very unlikely.
-    if (checkPrefix(_positivePrefix)) gotPositive = true;
-    if (checkPrefix(_negativePrefix)) gotNegative = true;
-
-    // The positive prefix might be a substring of the negative, in
-    // which case both would match.
-    if (gotPositive && gotNegative) {
-      if (_positivePrefix.length > _negativePrefix.length) {
-        gotNegative = false;
-      } else if (_negativePrefix.length > _positivePrefix.length) {
-        gotPositive = false;
-      }
-    }
-    if (skip) {
-      if (gotPositive) input.read(_positivePrefix.length);
-      if (gotNegative) input.read(_negativePrefix.length);
-    }
-  }
-
-  /// If the rest of our input is either the positive or negative suffix,
-  /// set [gotPositiveSuffix] or [gotNegativeSuffix] accordingly.
-  void checkSuffixes() {
-    var remainder = input.rest();
-    if (remainder == _positiveSuffix) gotPositiveSuffix = true;
-    if (remainder == _negativeSuffix) gotNegativeSuffix = true;
-  }
-
-  /// We've encountered a character that's not a digit. Go through our
-  /// replacement rules looking for how to handle it. If we see something
-  /// that's not a digit and doesn't have a replacement, then we're done
-  /// and the number is probably invalid.
-  void processNonDigit() {
-    // It might just be a prefix that we haven't skipped. We don't want to
-    // skip them initially because they might also be semantically meaningful,
-    // e.g. leading %. So we allow them through the loop, but only once.
-    var foundAnInterpretation = false;
-    if (input.index == 0 && !prefixesSkipped) {
-      prefixesSkipped = true;
-      checkPrefixes(skip: true);
-      foundAnInterpretation = true;
-    }
-
-    for (var key in replacements.keys) {
-      if (input.startsWith(key)) {
-        _normalized.write(replacements[key]());
-        input.read(key.length);
-        return;
-      }
-    }
-    // We haven't found either of these things, this seems invalid.
-    if (!foundAnInterpretation) {
-      done = true;
-    }
-  }
-
-  /// Parse [text] and return the resulting number. Throws [FormatException]
-  /// if we can't parse it.
-  num parse() {
-    if (text == symbols.NAN) return 0.0 / 0.0;
-    if (text == '$_positivePrefix${symbols.INFINITY}$_positiveSuffix') {
-      return 1.0 / 0.0;
-    }
-    if (text == '$_negativePrefix${symbols.INFINITY}$_negativeSuffix') {
-      return -1.0 / 0.0;
-    }
-
-    checkPrefixes();
-    var parsed = parseNumber(input);
-
-    if (gotPositive && !gotPositiveSuffix) invalidNumber();
-    if (gotNegative && !gotNegativeSuffix) invalidNumber();
-    if (!input.atEnd()) invalidNumber();
-
-    return parsed;
-  }
-
-  /// The number is invalid, throw a [FormatException].
-  void invalidNumber() =>
-      throw FormatException('Invalid Number: ${input.contents}');
-
-  /// Parse the number portion of the input, i.e. not any prefixes or suffixes,
-  /// and assuming NaN and Infinity are already handled.
-  num parseNumber(_Stream input) {
-    if (gotNegative) {
-      _normalized.write('-');
-    }
-    while (!done && !input.atEnd()) {
-      var digit = asDigit(input.peek());
-      if (digit != null) {
-        _normalized.writeCharCode(_zero + digit);
-        input.next();
-      } else {
-        processNonDigit();
-      }
-      checkSuffixes();
-    }
-
-    var normalizedText = _normalized.toString();
-    num parsed = int.tryParse(normalizedText);
-    parsed ??= double.parse(normalizedText);
-    return parsed / scale;
-  }
-}
-
-/// Private class that parses the numeric formatting pattern and sets the
-/// variables in [format] to appropriate values. Instances of this are
-/// transient and store parsing state in instance variables, so can only be used
-/// to parse a single pattern.
-class _NumberFormatParser {
-  /// The special characters in the pattern language. All others are treated
-  /// as literals.
-  static const _PATTERN_SEPARATOR = ';';
-  static const _QUOTE = "'";
-  static const _PATTERN_DIGIT = '#';
-  static const _PATTERN_ZERO_DIGIT = '0';
-  static const _PATTERN_GROUPING_SEPARATOR = ',';
-  static const _PATTERN_DECIMAL_SEPARATOR = '.';
-  static const _PATTERN_CURRENCY_SIGN = '\u00A4';
-  static const _PATTERN_PER_MILLE = '\u2030';
-  static const _PER_MILLE_SCALE = 1000;
-  static const _PATTERN_PERCENT = '%';
-  static const _PERCENT_SCALE = 100;
-  static const _PATTERN_EXPONENT = 'E';
-  static const _PATTERN_PLUS = '+';
-
-  /// The format whose state we are setting.
-  final NumberFormat format;
-
-  /// The pattern we are parsing.
-  final _StringIterator pattern;
-
-  /// We can be passed a specific currency symbol, regardless of the locale.
-  String currencySymbol;
-
-  /// We can be given a specific number of decimal places, overriding the
-  /// default.
-  final int decimalDigits;
-
-  /// Create a new [_NumberFormatParser] for a particular [NumberFormat] and
-  /// [input] pattern.
-  _NumberFormatParser(
-      this.format, input, this.currencySymbol, this.decimalDigits)
-      : pattern = _iterator(input) {
-    pattern.moveNext();
-  }
-
-  /// The [NumberSymbols] for the locale in which our [format] prints.
-  NumberSymbols get symbols => format.symbols;
-
-  /// Parse the input pattern and set the values.
-  void parse() {
-    format._positivePrefix = _parseAffix();
-    var trunk = _parseTrunk();
-    format._positiveSuffix = _parseAffix();
-    // If we have separate positive and negative patterns, now parse the
-    // the negative version.
-    if (pattern.current == _NumberFormatParser._PATTERN_SEPARATOR) {
-      pattern.moveNext();
-      format._negativePrefix = _parseAffix();
-      // Skip over the negative trunk, verifying that it's identical to the
-      // positive trunk.
-      for (var each in _iterable(trunk)) {
-        if (pattern.current != each && pattern.current != null) {
-          throw FormatException(
-              'Positive and negative trunks must be the same', trunk);
-        }
-        pattern.moveNext();
-      }
-      format._negativeSuffix = _parseAffix();
-    } else {
-      // If no negative affix is specified, they share the same positive affix.
-      format._negativePrefix = format._negativePrefix + format._positivePrefix;
-      format._negativeSuffix = format._positiveSuffix + format._negativeSuffix;
-    }
-  }
-
-  /// Variable used in parsing prefixes and suffixes to keep track of
-  /// whether or not we are in a quoted region.
-  bool inQuote = false;
-
-  /// Parse a prefix or suffix and return the prefix/suffix string. Note that
-  /// this also may modify the state of [format].
-  String _parseAffix() {
-    var affix = StringBuffer();
-    inQuote = false;
-    while (parseCharacterAffix(affix) && pattern.moveNext()) {}
-    return affix.toString();
-  }
-
-  /// Parse an individual character as part of a prefix or suffix.  Return true
-  /// if we should continue to look for more affix characters, and false if
-  /// we have reached the end.
-  bool parseCharacterAffix(StringBuffer affix) {
-    var ch = pattern.current;
-    if (ch == null) return false;
-    if (ch == _QUOTE) {
-      if (pattern.peek == _QUOTE) {
-        pattern.moveNext();
-        affix.write(_QUOTE); // 'don''t'
-      } else {
-        inQuote = !inQuote;
-      }
-      return true;
-    }
-
-    if (inQuote) {
-      affix.write(ch);
-    } else {
-      switch (ch) {
-        case _PATTERN_DIGIT:
-        case _PATTERN_ZERO_DIGIT:
-        case _PATTERN_GROUPING_SEPARATOR:
-        case _PATTERN_DECIMAL_SEPARATOR:
-        case _PATTERN_SEPARATOR:
-          return false;
-        case _PATTERN_CURRENCY_SIGN:
-          // TODO(alanknight): Handle the local/global/portable currency signs
-          affix.write(currencySymbol);
-          break;
-        case _PATTERN_PERCENT:
-          if (format._multiplier != 1 && format._multiplier != _PERCENT_SCALE) {
-            throw FormatException('Too many percent/permill', format);
-          }
-          format._multiplier = _PERCENT_SCALE;
-          affix.write(symbols.PERCENT);
-          break;
-        case _PATTERN_PER_MILLE:
-          if (format._multiplier != 1 &&
-              format._multiplier != _PER_MILLE_SCALE) {
-            throw FormatException('Too many percent/permill', format);
-          }
-          format._multiplier = _PER_MILLE_SCALE;
-          affix.write(symbols.PERMILL);
-          break;
-        default:
-          affix.write(ch);
-      }
-    }
-    return true;
-  }
-
-  /// Variables used in [_parseTrunk] and [parseTrunkCharacter].
-  var decimalPos = -1;
-  var digitLeftCount = 0;
-  var zeroDigitCount = 0;
-  var digitRightCount = 0;
-  var groupingCount = -1;
-
-  /// Parse the "trunk" portion of the pattern, the piece that doesn't include
-  /// positive or negative prefixes or suffixes.
-  String _parseTrunk() {
-    var loop = true;
-    var trunk = StringBuffer();
-    while (pattern.current != null && loop) {
-      loop = parseTrunkCharacter(trunk);
-    }
-
-    if (zeroDigitCount == 0 && digitLeftCount > 0 && decimalPos >= 0) {
-      // Handle '###.###' and '###.' and '.###'
-      // Handle '.###'
-      var n = decimalPos == 0 ? 1 : decimalPos;
-      digitRightCount = digitLeftCount - n;
-      digitLeftCount = n - 1;
-      zeroDigitCount = 1;
-    }
-
-    // Do syntax checking on the digits.
-    if (decimalPos < 0 && digitRightCount > 0 ||
-        decimalPos >= 0 &&
-            (decimalPos < digitLeftCount ||
-                decimalPos > digitLeftCount + zeroDigitCount) ||
-        groupingCount == 0) {
-      throw FormatException('Malformed pattern "${pattern.input}"');
-    }
-    var totalDigits = digitLeftCount + zeroDigitCount + digitRightCount;
-
-    format.maximumFractionDigits =
-        decimalPos >= 0 ? totalDigits - decimalPos : 0;
-    if (decimalPos >= 0) {
-      format.minimumFractionDigits =
-          digitLeftCount + zeroDigitCount - decimalPos;
-      if (format.minimumFractionDigits < 0) {
-        format.minimumFractionDigits = 0;
-      }
-    }
-
-    // The effectiveDecimalPos is the position the decimal is at or would be at
-    // if there is no decimal. Note that if decimalPos<0, then digitTotalCount
-    // == digitLeftCount + zeroDigitCount.
-    var effectiveDecimalPos = decimalPos >= 0 ? decimalPos : totalDigits;
-    format.minimumIntegerDigits = effectiveDecimalPos - digitLeftCount;
-    if (format._useExponentialNotation) {
-      format.maximumIntegerDigits =
-          digitLeftCount + format.minimumIntegerDigits;
-
-      // In exponential display, we need to at least show something.
-      if (format.maximumFractionDigits == 0 &&
-          format.minimumIntegerDigits == 0) {
-        format.minimumIntegerDigits = 1;
-      }
-    }
-
-    format._finalGroupingSize = max(0, groupingCount);
-    if (!format._groupingSizeSetExplicitly) {
-      format._groupingSize = format._finalGroupingSize;
-    }
-    format._decimalSeparatorAlwaysShown =
-        decimalPos == 0 || decimalPos == totalDigits;
-
-    return trunk.toString();
-  }
-
-  /// Parse an individual character of the trunk. Return true if we should
-  /// continue to look for additional trunk characters or false if we have
-  /// reached the end.
-  bool parseTrunkCharacter(trunk) {
-    var ch = pattern.current;
-    switch (ch) {
-      case _PATTERN_DIGIT:
-        if (zeroDigitCount > 0) {
-          digitRightCount++;
-        } else {
-          digitLeftCount++;
-        }
-        if (groupingCount >= 0 && decimalPos < 0) {
-          groupingCount++;
-        }
-        break;
-      case _PATTERN_ZERO_DIGIT:
-        if (digitRightCount > 0) {
-          throw FormatException('Unexpected "0" in pattern "${pattern.input}');
-        }
-        zeroDigitCount++;
-        if (groupingCount >= 0 && decimalPos < 0) {
-          groupingCount++;
-        }
-        break;
-      case _PATTERN_GROUPING_SEPARATOR:
-        if (groupingCount > 0) {
-          format._groupingSizeSetExplicitly = true;
-          format._groupingSize = groupingCount;
-        }
-        groupingCount = 0;
-        break;
-      case _PATTERN_DECIMAL_SEPARATOR:
-        if (decimalPos >= 0) {
-          throw FormatException(
-              'Multiple decimal separators in pattern "$pattern"');
-        }
-        decimalPos = digitLeftCount + zeroDigitCount + digitRightCount;
-        break;
-      case _PATTERN_EXPONENT:
-        trunk.write(ch);
-        if (format._useExponentialNotation) {
-          throw FormatException(
-              'Multiple exponential symbols in pattern "$pattern"');
-        }
-        format._useExponentialNotation = true;
-        format.minimumExponentDigits = 0;
-
-        // exponent pattern can have a optional '+'.
-        pattern.moveNext();
-        var nextChar = pattern.current;
-        if (nextChar == _PATTERN_PLUS) {
-          trunk.write(pattern.current);
-          pattern.moveNext();
-          format._useSignForPositiveExponent = true;
-        }
-
-        // Use lookahead to parse out the exponential part
-        // of the pattern, then jump into phase 2.
-        while (pattern.current == _PATTERN_ZERO_DIGIT) {
-          trunk.write(pattern.current);
-          pattern.moveNext();
-          format.minimumExponentDigits++;
-        }
-
-        if ((digitLeftCount + zeroDigitCount) < 1 ||
-            format.minimumExponentDigits < 1) {
-          throw FormatException('Malformed exponential pattern "$pattern"');
-        }
-        return false;
-      default:
-        return false;
-    }
-    trunk.write(ch);
-    pattern.moveNext();
-    return true;
-  }
-}
-
-/// Returns an [Iterable] on the string as a list of substrings.
-Iterable<String> _iterable(String s) => _StringIterable(s);
-
-/// Return an iterator on the string as a list of substrings.
-Iterator<String> _iterator(String s) => _StringIterator(s);
-
-// TODO(nweiz): remove this when issue 3780 is fixed.
-/// Provides an Iterable that wraps [_iterator] so it can be used in a `for`
-/// loop.
-class _StringIterable extends IterableBase<String> {
-  final Iterator<String> iterator;
-
-  _StringIterable(String s) : iterator = _iterator(s);
-}
-
-/// Provides an iterator over a string as a list of substrings, and also
-/// gives us a lookahead of one via the [peek] method.
-class _StringIterator implements Iterator<String> {
-  final String input;
-  int nextIndex = 0;
-  String _current;
-
-  _StringIterator(input) : input = _validate(input);
-
-  String get current => _current;
-
-  bool moveNext() {
-    if (nextIndex >= input.length) {
-      _current = null;
-      return false;
-    }
-    _current = input[nextIndex++];
-    return true;
-  }
-
-  String get peek => nextIndex >= input.length ? null : input[nextIndex];
-
-  Iterator<String> get iterator => this;
-
-  static String _validate(input) {
-    if (input is! String) throw ArgumentError(input);
-    return input;
-  }
-}
-
-/// Used primarily for currency formatting, this number-like class stores
-/// millionths of a currency unit, typically as an Int64.
-///
-/// It supports no operations other than being used for Intl number formatting.
-abstract class MicroMoney {
-  factory MicroMoney(micros) => _MicroMoney(micros);
-}
-
-/// Used primarily for currency formatting, this stores millionths of a
-/// currency unit, typically as an Int64.
-///
-/// This private class provides the operations needed by the formatting code.
-class _MicroMoney implements MicroMoney {
-  final dynamic _micros;
-  _MicroMoney(this._micros);
-  static const _multiplier = 1000000;
-
-  dynamic get _integerPart => _micros ~/ _multiplier;
-  int get _fractionPart => (this - _integerPart)._micros.toInt().abs();
-
-  bool get isNegative => _micros.isNegative;
-
-  _MicroMoney abs() => isNegative ? _MicroMoney(_micros.abs()) : this;
-
-  // Note that if this is done in a general way there's a risk of integer
-  // overflow on JS when multiplying out the [other] parameter, which may be
-  // an Int64. In formatting we only ever subtract out our own integer part.
-  _MicroMoney operator -(other) {
-    if (other is _MicroMoney) return _MicroMoney(_micros - other._micros);
-    return _MicroMoney(_micros - (other * _multiplier));
-  }
-
-  _MicroMoney operator +(other) {
-    if (other is _MicroMoney) return _MicroMoney(_micros + other._micros);
-    return _MicroMoney(_micros + (other * _multiplier));
-  }
-
-  _MicroMoney operator ~/(divisor) {
-    if (divisor is! int) {
-      throw ArgumentError.value(
-          divisor, 'divisor', '_MicroMoney ~/ only supports int arguments.');
-    }
-    return _MicroMoney((_integerPart ~/ divisor) * _multiplier);
-  }
-
-  _MicroMoney operator *(other) {
-    if (other is! int) {
-      throw ArgumentError.value(
-          other, 'other', '_MicroMoney * only supports int arguments.');
-    }
-    return _MicroMoney(
-        (_integerPart * other) * _multiplier + (_fractionPart * other));
-  }
-
-  /// Note that this only really supports remainder from an int,
-  /// not division by another MicroMoney
-  _MicroMoney remainder(other) {
-    if (other is! int) {
-      throw ArgumentError.value(
-          other, 'other', '_MicroMoney.remainder only supports int arguments.');
-    }
-    return _MicroMoney(_micros.remainder(other * _multiplier));
-  }
-
-  double toDouble() => _micros.toDouble() / _multiplier;
-
-  int toInt() => _integerPart.toInt();
-
-  String toString() {
-    var beforeDecimal = '$_integerPart';
-    var decimalPart = '';
-    var fractionPart = _fractionPart;
-    if (fractionPart != 0) {
-      decimalPart = '.$fractionPart';
-    }
-    return '$beforeDecimal$decimalPart';
-  }
-}
+final _ln10 = log(10);
