@@ -31,11 +31,11 @@ PluralCase _default_rule() => OTHER;
 
 /// This must be called before evaluating a new rule, because we're using
 /// library-global state to both keep the rules terse and minimize space.
-void startRuleEvaluation(num howMany, [int precision = 0]) {
+void startRuleEvaluation(num howMany, [int? precision = 0]) {
   _n = howMany;
   _precision = precision;
   _i = _n.round();
-  _updateVF(_n, _precision);
+  _updateVF(_n);
   _updateWT(_f, _v);
 }
 
@@ -46,21 +46,21 @@ void startRuleEvaluation(num howMany, [int precision = 0]) {
 // not introduce a subclass per locale or have instance tear-offs which
 // we can't cache. This is fine as long as these methods aren't async, which
 // they should never be.
-num _n;
+num _n = 0;
 
 /// The integer part of [_n]
-int _i;
-int _precision;
+int _i = 0;
+int? _precision;
 
 /// Returns the number of digits in the fractional part of a number
 /// (3.1416 => 4)
 ///
-/// Takes the item count [n] and a [precision].
+/// Takes the item count [n] and uses [_precision].
 /// That's because a just looking at the value of a number is not enough to
 /// decide the plural form. For example "1 dollar" vs "1.00 dollars", the
 /// value is 1, but how it is formatted also matters.
-int _decimals(num n, int precision) {
-  var str = _precision == null ? '$n' : n.toStringAsFixed(precision);
+int _decimals(num n) {
+  var str = _precision == null ? '$n' : n.toStringAsFixed(_precision!);
   var result = str.indexOf('.');
   return (result == -1) ? 0 : str.length - result - 1;
 }
@@ -70,10 +70,10 @@ int _decimals(num n, int precision) {
 /// The short names for parameters / return match the CLDR syntax and UTS #35
 ///     (https://unicode.org/reports/tr35/tr35-numbers.html#Plural_rules_syntax)
 /// Takes the item count [n] and a [precision].
-void _updateVF(num n, int precision) {
+void _updateVF(num n) {
   var defaultDigits = 3;
 
-  _v = precision ?? math.min(_decimals(n, precision), defaultDigits);
+  _v = _precision ?? math.min(_decimals(n), defaultDigits);
 
   var base = math.pow(10, _v) as int;
   _f = (n * base).floor() % base;
