@@ -16,7 +16,7 @@ library date_time_patterns;
 /// Returns a Map from locale names to another Map that goes from skeletons
 /// to the locale-specific formatting patterns.
 /// Internal use only. Call initializeDateFormatting instead.
-Map<dynamic, dynamic> dateTimePatternMap() => const {
+Map<String, Map<String, String>> dateTimePatternMap() => const {
       /// Extended set of localized date/time patterns for locale af.
       'af': const {
         'd': 'd', // DAY
