@@ -2,7 +2,10 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-part of intl;
+import 'dart:convert';
+
+import 'bidi.dart';
+import 'text_direction.dart';
 
 // Suppress naming issues as changing them would be breaking.
 // ignore_for_file: non_constant_identifier_names
@@ -67,13 +70,13 @@ class BidiFormatter {
   /// should always use a `span` tag, even when the input directionality is
   /// neutral or matches the context, so that the DOM structure of the output
   /// does not depend on the combination of directionalities.
-  BidiFormatter.LTR([alwaysSpan = false])
+  BidiFormatter.LTR([bool alwaysSpan = false])
       : contextDirection = TextDirection.LTR,
         _alwaysSpan = alwaysSpan;
-  BidiFormatter.RTL([alwaysSpan = false])
+  BidiFormatter.RTL([bool alwaysSpan = false])
       : contextDirection = TextDirection.RTL,
         _alwaysSpan = alwaysSpan;
-  BidiFormatter.UNKNOWN([alwaysSpan = false])
+  BidiFormatter.UNKNOWN([bool alwaysSpan = false])
       : contextDirection = TextDirection.UNKNOWN,
         _alwaysSpan = alwaysSpan;
 
@@ -96,7 +99,7 @@ class BidiFormatter {
   /// a trailing unicode BiDi mark matching the context directionality is
   /// appended (LRM or RLM). If [isHtml] is false, we HTML-escape the [text].
   String wrapWithSpan(String text,
-      {bool isHtml = false, bool resetDir = true, TextDirection direction}) {
+      {bool isHtml = false, bool resetDir = true, TextDirection? direction}) {
     direction ??= estimateDirection(text, isHtml: isHtml);
     String result;
     if (!isHtml) text = const HtmlEscape().convert(text);
@@ -131,7 +134,7 @@ class BidiFormatter {
   /// [isHtml]. [isHtml] is used to designate if the text contains HTML (escaped
   /// or unescaped).
   String wrapWithUnicode(String text,
-      {bool isHtml = false, bool resetDir = true, TextDirection direction}) {
+      {bool isHtml = false, bool resetDir = true, TextDirection? direction}) {
     direction ??= estimateDirection(text, isHtml: isHtml);
     var result = text;
     if (contextDirection.isDirectionChange(direction)) {
